requirements.txt
flask
py-tlsh
ssdeep
python-magic
pefile
pyelftools
PyPDF2
python-docx
pymupdf